backend/requirements.txt
python-multipart
email-validator
requests
celery
passlib[bcrypt]
tenacity
raven
gunicorn
alembic
python-jose[cryptography]
boto3
unidecode
fastapi-mail
uvicorn
Jinja2
fastapi-users[sqlalchemy]
greenlet
psycopg2cffi
sqlmodel
networkx
pymgclient
fastapi-users-db-sqlalchemy
makefun
fastapi-users-db-beanie